updated media organizer to more folders to trash when done with them.

+++ experiments/media-organizer.py
@@ -2,8 +2,13 @@
 
 import os, sys, pdb, re, shutil
 import datetime
+import gntp.notifier
 import transmissionrpc
 from transmissionrpc.error import TransmissionError
+
+class FlagFile:
+    REMOVE_FOLDER = '.removefolder'
+    UNRARED = '.unrared'
 
 class Organizer(object):
     def __init__(self):
@@ -12,30 +17,48 @@ class Organizer(object):
         self.DOWNLOAD_FOLDER = "/Users/jasonwaters/Downloads"
         self.TV_FOLDER = "/Volumes/Blue/TV"
         self.MOVIE_FOLDER = "/Volumes/Blue/Movies"
+        self.TRASH = "/Users/jasonwaters/.Trash"
 
         self.TRANSMISSION_HOST = "localhost"
         self.TRANSMISSION_PORT = 9091
-        self.TRANSMISSION_USER = "admin"
-        self.TRANSMISSION_PASSWORD = "user"
+        self.TRANSMISSION_USER = ""
+        self.TRANSMISSION_PASSWORD = ""
 
-        self.PATTERN_EPISODE = re.compile(".*(([sS]\d+[eE]\d+)|(\d+x\d+)).*")
+        self.PATTERN_EPISODE = re.compile(".*((([sS]\d{1,2}[eE]\d{1,2})|(\d+x\d+))|(\.\d{3}\.)).*")
         self.PATTERN_VIDEO = re.compile("(^.+\.(avi|mp4|mkv)$)")
         self.PATTERN_RAR = re.compile("^.+\.(rar|r\d+)$")
 
-        self.mark_file_name = '.unrared'
+        self.growl = gntp.notifier.GrowlNotifier(
+            applicationName = "Automatic Media Organizer",
+            notifications = ["New Updates","New Messages"],
+            defaultNotifications = ["New Messages"],
+        )
+
+        self.growl.register()
+
         self.extensions_unrar = ['.rar', '.r01']        # List of extensions for auto-extract to look for
         self.supported_filetypes = []                   # Filled by extensions_list function
         self.extensions_list()
 
-        self.log("="*40)
         self.unrar_check()
+
         # Check that the download directory parameters is actually a directory
-        self.removeFinishedTorrents()
+        self.removeFinishedTorrents() #from transmission
+
         self.traverse_directories()
 
     def log(self, message):
         now = datetime.datetime.now()
-        print "%s | %s\n" % (now.strftime("%Y-%m-%d %H:%M:%S"), message)
+        print "%s (%s)\n" % (message, now.strftime("%Y-%m-%d %H:%M:%S"))
+
+    def notify(self, notifyTitle, notifyMessage):
+        self.growl.notify(
+            noteType = "New Messages",
+            title = notifyTitle,
+            description = notifyMessage,
+            sticky = False,
+            priority = 1,
+        )
 
     '''Creates the list of extensions supported by the script'''
     def extensions_list(self):
@@ -65,7 +88,8 @@ class Organizer(object):
     def unrar_check(self):
         self.unrar_exe()
         self.unrar_path = self.find_unrar()
-        if self.unrar_path != False:
+
+        if self.unrar_path:
             self.unrar_exe = os.path.join(self.unrar_path, self.unrar_name)
         else:
             self.log('Error: ' + self.unrar_name + ' not found in the system path')
@@ -77,6 +101,7 @@ class Organizer(object):
         for dirname, dirnames, filenames in os.walk(self.DOWNLOAD_FOLDER):
             self.scan_for_archives(dirname)
             self.scan_for_videos(dirname)
+            self.clean_up(dirname)
 
     '''Check for rar files in each directory'''
     def scan_for_archives(self, dir):
@@ -88,14 +113,11 @@ class Organizer(object):
             for ext in self.supported_filetypes:
                 if filename.endswith(ext):
                     # If a .rar archive is found, check to see if it has been extracted previously
-                    file_unrared = os.path.exists(os.path.join(dir, self.mark_file_name))
+                    file_unrared = os.path.exists(os.path.join(dir, FlagFile.UNRARED))
                     if not file_unrared:
                         self.log("Need to extract: " + filename)
                         # Start extracting file
                         self.start_unrar(dir, filename)
-                    else:
-                        self.log('Skipping archive ' + filename)
-                        self.delete_rars(dir)
                     # .rar was found, dont need to search for .r01
                     break
 
@@ -114,17 +136,35 @@ class Organizer(object):
                 else:
                     shutil.move(filePath, self.MOVIE_FOLDER)
                     folder = self.MOVIE_FOLDER
+                self.mark_dir(dir, FlagFile.REMOVE_FOLDER)
 
-                self.log("Moved %s to '%s' folder" % (filename, folder))
+                self.log("Moved '%s' from '%s' to '%s'." % (filename, filePath, folder))
+                self.notify("Moved '%s'." % filename, "'%s' was moved to '%s'" % (filename, folder))
+
+    def clean_up(self, dir):
+        unrared = os.path.exists(os.path.join(dir, FlagFile.UNRARED))
+        removeFolder = os.path.exists(os.path.join(dir, FlagFile.REMOVE_FOLDER))
+
+        if unrared:
+            self.delete_rars(dir)
+
+        if removeFolder:
+            self.trashFolder(dir)
 
     def delete_rars(self, dir):
-        self.log("Deleting RARs...")
-        dir_listing = os.listdir(dir)
+        if dir is not self.DOWNLOAD_FOLDER:
+            self.log("Deleting RARs (%s)..." % dir)
+            dir_listing = os.listdir(dir)
 
-        for filename in dir_listing:
-            if self.isRar(filename):
-                self.log("deleted %s" % filename)
-                os.remove(os.path.join(dir,filename))
+            for filename in dir_listing:
+                if self.isRar(filename):
+                    self.log("deleted %s" % filename)
+                    os.remove(os.path.join(dir,filename))
+
+    def trashFolder(self, dir):
+        if dir is not self.DOWNLOAD_FOLDER:
+            self.log("Deleting Folder...")
+            shutil.move(dir, self.TRASH)
 
     '''Extract a rar archive'''
     def start_unrar(self, dir, archive_name):
@@ -143,15 +183,15 @@ class Organizer(object):
             exit()
 
         # Sucessfully extracted archive, mark the dir with a hidden file
-        self.mark_dir(dir)
+        self.mark_dir(dir, FlagFile.UNRARED)
         self.delete_rars(dir)
 
     '''Creates a hidden file so the same archives will not be extracted again'''
-    def mark_dir(self, dir):
-        mark_file = os.path.join(dir, self.mark_file_name)
+    def mark_dir(self, dir, mark_file_name):
+        mark_file = os.path.join(dir, mark_file_name)
         f = open(mark_file,'w')
         f.close()
-        self.log(self.mark_file_name + ' file created')
+        self.log(mark_file_name + ' file created')
 
     def isRar(self, name):
         return self.PATTERN_RAR.match(name.lower()) is not None
@@ -160,7 +200,7 @@ class Organizer(object):
         return self.PATTERN_EPISODE.match(name.lower()) is not None
 
     def isValidVideoFile(self, name):
-        return self.PATTERN_VIDEO.match(name.lower()) is not None and name.find('sample') == -1
+        return self.PATTERN_VIDEO.match(name.lower()) is not None and name.lower().find('sample') == -1
 
     def removeFinishedTorrents(self):
         try:
@@ -173,7 +213,8 @@ class Organizer(object):
 
         for tid, torrent in torrents.iteritems():
             if torrent.progress == 100:
-                self.log('Removed   %s.' % torrent.name)
+                self.log("Removed '%s'." % torrent.name)
+                self.notify("Removed '%s'." % torrent.name, "'%s' was removed from Transmission." % torrent.name)
                 client.remove(torrent.hashString, delete_data=False)
